requirement.txt
streamlit
python-dotenv
google-generativeai
requests
pandas
bcrypt
fastapi
uvicorn